:repeat: Rename nav.js to navbar.js for clarity
Updated the file name and import to improve readability and consistency. This change ensures the file name better reflects its content and function.
