+++ assets/src/js/frontend.ts
@@ -2,7 +2,7 @@ import "../css/tailwind.css";
 import "../css/frontend.css";
 import "./accordion";
 
-import { initPrimaryNav } from "./nav";
+import { initPrimaryNav } from "./navbar";
 
 document.addEventListener("DOMContentLoaded", () => {
 	initPrimaryNav();
